PERF-1952 POC: Add scavenge method to tcmalloc central cache.
If application has deallocated all objects of a given span, some
objects may still be held in the free lists of central cache or
thread cache. This patch tracks when objects are returned to the
central cache and actively scavenges them to the span if this
would cause the span to be returned to the page heap.